requirements.txt
matplotlib-3.2.1
numpy-1.18.5
pandas-1.15.0
seaborn-0.10.1
